Fix mypy errors blocking CI (typed test fixtures + build_pq overloads)
CI also runs mypy in strict-ish mode; the migration commit introduced
several errors:

- ``build_pq`` returned a union type (``tuple[A,B,C,D] | tuple[A,B,C,D,E]``)
  driven by ``return_metadata: bool``. mypy can't narrow on bool, so every
  4-tuple unpack at the call site failed. Switched to ``@overload`` pair
  with ``Literal[False]`` / ``Literal[True]`` so the static return type
  matches the runtime shape.
- ``back_substitute`` indexed a ``dict[str, object]`` metadata dict and
  passed values to ``int(...)`` / tuple-unpacking. ``cast()`` each read
  to its known runtime type.
- ``solve_x2_roots_mobius``: explicit type annotations on ``real_roots`` /
  ``real_eigvecs`` were duplicated in two branches of the same scope; mypy
  flagged the second as redefinition. Removed the second annotation
  (mypy infers from the first).
- ``poe_to_dh``: ``np.linalg.solve`` returns ``floating[Any]``, not
  ``float64``. ``.astype(np.float64)`` makes the dtype explicit.
- ``test_refinement.py`` helpers were untyped; mypy treats untyped helpers
  as untyped-call sites in the typed test bodies. Added ``KinBody``,
  ``NDArray[np.float64]`` annotations on ``_fk_poe`` and the ``ur5_kb``
  fixture; sed-applied ``ur5_kb: KinBody`` to all 8 test signatures.
- ``test_puma_cross_validation._solution_sets_equal`` migrated to accept
  ``Solution | NDArray`` via ``Any`` parameterization. Tightened the
  ``_q`` helper return path so mypy doesn't see ``Any`` leaking.
- ``scipy.linalg.eig`` import: ``# type: ignore[import-untyped]`` (no
  scipy-stubs in the runtime env, and we already gracefully ImportError-
  fall through if scipy is missing).

uv run mypy src tests: Success: no issues found in 75 source files
uv run ruff check + format --check: All checks passed
uv run pytest tests/test_refinement.py tests/test_solution.py
tests/test_ikgeo_general_6r.py tests/test_jaco2_general_6r.py: 19/19

+++ src/ssik/kinematics/poe_to_dh.py
@@ -297,7 +297,7 @@ def poe_to_dh(kb: KinBody) -> DhWithOffset:
     t_dh_end[:3, 2] = z_n
     t_dh_end[:3, 3] = origins[n]
     # We need: t_dh_end @ t_post = t_home  ->  t_post = t_dh_end^{-1} @ t_home
-    t_post = np.linalg.solve(t_dh_end, t_home)
+    t_post = np.linalg.solve(t_dh_end, t_home).astype(np.float64)
 
     return DhWithOffset(
         alpha=alpha,
